Changed page number to be int everywhere

+++ app/src/main/java/com/example/pecodetask/features/pageContainer/domain/model/PageItem.kt
@@ -1,5 +1,5 @@
 package com.example.pecodetask.features.pageContainer.domain.model
 
 data class PageItem(
-    val pageNumber: Long
+    val pageNumber: Int
 )
