+++ app/src/main/java/com/example/pecodetask/features/pageContent/domain/model/NotificationData.kt
@@ -2,5 +2,5 @@ package com.example.pecodetask.features.pageContent.domain.model
 
 data class NotificationData(
     val id: Int,
-    val pageNumber: Long
+    val pageNumber: Int
 )
